pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flux2-litekit"
version = "0.1.0"
description = "Lightweight training and inference toolkit based on FLUX.2."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
]
dependencies = [
  "accelerate",
  "diffusers",
  "peft",
  "pillow",
  "pyyaml",
  "safetensors",
  "torch",
  "transformers",
  "tqdm",
]

[tool.setuptools]
packages = ["flux2_litekit"]
